+++ bundle/jsonc-full.ts
@@ -1,0 +1,4 @@
+import {registerJSONC} from '../src/index';
+export {CodeMirror6} from '../src/index';
+
+registerJSONC();