+++ bundle/jsonc.ts
@@ -1,0 +1,4 @@
+import {registerJSONCCore} from '../src/index';
+export {CodeMirror6} from '../src/index';
+
+registerJSONCCore();